extract dataclasses for fuels2018
